+++ example/lib/examples/example10.dart
@@ -1,16 +1,17 @@
 import 'dart:math';
 
 import 'package:easy_sticky_header/easy_sticky_header.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-class Example9 extends StatefulWidget {
-  const Example9({Key? key}) : super(key: key);
+class Example10 extends StatefulWidget {
+  const Example10({Key? key}) : super(key: key);
 
   @override
-  State<Example9> createState() => _Example9State();
+  State<Example10> createState() => _Example10State();
 }
 
-class _Example9State extends State<Example9> {
+class _Example10State extends State<Example10> {
   late final StickyHeaderController _controller;
 
   @override
@@ -32,8 +33,7 @@ class _Example9State extends State<Example9> {
       appBar: AppBar(
         shadowColor: Colors.transparent,
         backgroundColor: Colors.black,
-        title:
-            const Text('Scrolls to the top after the header widget is tapped'),
+        title: const Text('Jumps to the header of the specified index'),
       ),
       body: StickyHeader(
         controller: _controller,
@@ -43,28 +43,21 @@ class _Example9State extends State<Example9> {
           ),
           itemCount: 100,
           itemBuilder: (context, index) {
-            if (index % 6 == 0) {
+            if (index % 3 == 0) {
               return StickyContainerWidget(
                 index: index,
-                child: GestureDetector(
-                  onTap: () {
-                    // Jumps without animation.
-                    // _controller.jumpTo(index);
-                    _controller.animateTo(index);
-                  },
-                  child: Container(
-                    color: Color.fromRGBO(Random().nextInt(256),
-                        Random().nextInt(256), Random().nextInt(256), 1),
-                    padding: const EdgeInsets.only(left: 16.0),
-                    alignment: Alignment.centerLeft,
-                    width: double.infinity,
-                    height: 50,
-                    child: Text(
-                      'Header #$index',
-                      style: const TextStyle(
-                        color: Colors.white,
-                        fontSize: 16,
-                      ),
+                child: Container(
+                  color: Color.fromRGBO(Random().nextInt(256),
+                      Random().nextInt(256), Random().nextInt(256), 1),
+                  padding: const EdgeInsets.only(left: 16.0),
+                  alignment: Alignment.centerLeft,
+                  width: double.infinity,
+                  height: 50,
+                  child: Text(
+                    'Header #$index',
+                    style: const TextStyle(
+                      color: Colors.white,
+                      fontSize: 16,
                     ),
                   ),
                 ),
@@ -95,6 +88,23 @@ class _Example9State extends State<Example9> {
               ],
             );
           },
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          /// Note that this jump operation is cache based. If the header widget
+          /// information of the specified index is not in the cache, no
+          /// operation will be performed and false will be returned.
+          if (kDebugMode) {
+            print('success: ${_controller.jumpTo(6)}');
+          }
+        },
+        child: const Text(
+          'Jump',
+          style: TextStyle(
+            color: Colors.white,
+            fontSize: 10,
+          ),
         ),
       ),
     );