+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.1
 * Support for earlier version of dart sdk.
+* Update README and examples.
 
 ## 1.0.0
 * Initial version.